# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lead Management CRM API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface AdminUser {
  id: number;
  email: string;
  name: string;
}

export interface LoginResponse {
  token: string;
  user: AdminUser;
}

export type LeadSource = (typeof LeadSource)[keyof typeof LeadSource];

export const LeadSource = {
  Website: "Website",
  Social_Media: "Social Media",
  Referral: "Referral",
  Cold_Call: "Cold Call",
  Email_Campaign: "Email Campaign",
  Other: "Other",
} as const;

export type LeadStatus = (typeof LeadStatus)[keyof typeof LeadStatus];

export const LeadStatus = {
  New: "New",
  Contacted: "Contacted",
  Converted: "Converted",
} as const;

export interface Lead {
  id: number;
  name: string;
  email: string;
  phone: string;
  source: LeadSource;
  status: LeadStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateLeadBodySource =
  (typeof CreateLeadBodySource)[keyof typeof CreateLeadBodySource];

export const CreateLeadBodySource = {
  Website: "Website",
  Social_Media: "Social Media",
  Referral: "Referral",
  Cold_Call: "Cold Call",
  Email_Campaign: "Email Campaign",
  Other: "Other",
} as const;

export type CreateLeadBodyStatus =
  (typeof CreateLeadBodyStatus)[keyof typeof CreateLeadBodyStatus];

export const CreateLeadBodyStatus = {
  New: "New",
  Contacted: "Contacted",
  Converted: "Converted",
} as const;

export interface CreateLeadBody {
  name: string;
  email: string;
  phone: string;
  source: CreateLeadBodySource;
  status: CreateLeadBodyStatus;
  /** @nullable */
  notes?: string | null;
}

export type UpdateLeadBodySource =
  (typeof UpdateLeadBodySource)[keyof typeof UpdateLeadBodySource];

export const UpdateLeadBodySource = {
  Website: "Website",
  Social_Media: "Social Media",
  Referral: "Referral",
  Cold_Call: "Cold Call",
  Email_Campaign: "Email Campaign",
  Other: "Other",
} as const;

export type UpdateLeadBodyStatus =
  (typeof UpdateLeadBodyStatus)[keyof typeof UpdateLeadBodyStatus];

export const UpdateLeadBodyStatus = {
  New: "New",
  Contacted: "Contacted",
  Converted: "Converted",
} as const;

export interface UpdateLeadBody {
  name?: string;
  email?: string;
  phone?: string;
  source?: UpdateLeadBodySource;
  status?: UpdateLeadBodyStatus;
  /** @nullable */
  notes?: string | null;
}

export type UpdateLeadStatusBodyStatus =
  (typeof UpdateLeadStatusBodyStatus)[keyof typeof UpdateLeadStatusBodyStatus];

export const UpdateLeadStatusBodyStatus = {
  New: "New",
  Contacted: "Contacted",
  Converted: "Converted",
} as const;

export interface UpdateLeadStatusBody {
  status: UpdateLeadStatusBodyStatus;
}

export interface LeadsListResponse {
  leads: Lead[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface DashboardStats {
  totalLeads: number;
  newLeads: number;
  contactedLeads: number;
  convertedLeads: number;
  conversionRate: number;
  leadsThisMonth: number;
}

export interface SourceBreakdown {
  source: string;
  count: number;
}

export type GetLeadsParams = {
  /**
   * Search by name or email
   */
  search?: string;
  /**
   * Filter by status
   */
  status?: GetLeadsStatus;
  /**
   * Sort by date
   */
  sort?: GetLeadsSort;
  /**
   * Page number
   */
  page?: number;
  /**
   * Items per page
   */
  limit?: number;
};

export type GetLeadsStatus =
  (typeof GetLeadsStatus)[keyof typeof GetLeadsStatus];

export const GetLeadsStatus = {
  New: "New",
  Contacted: "Contacted",
  Converted: "Converted",
} as const;

export type GetLeadsSort = (typeof GetLeadsSort)[keyof typeof GetLeadsSort];

export const GetLeadsSort = {
  newest: "newest",
  oldest: "oldest",
} as const;