# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Lead Management CRM API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: leads
    description: Lead management operations
  - name: dashboard
    description: Dashboard and analytics operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Admin login
      description: Authenticate with email and password, returns JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      description: Returns authenticated user info
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /leads:
    get:
      operationId: getLeads
      tags: [leads]
      summary: List all leads
      description: Returns a list of leads with optional search and filter
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Search by name or email
        - name: status
          in: query
          schema:
            type: string
            enum: [New, Contacted, Converted]
          description: Filter by status
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, oldest]
          description: Sort by date
        - name: page
          in: query
          schema:
            type: integer
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
          description: Items per page
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadsListResponse"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a new lead
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadBody"
      responses:
        "201":
          description: Lead created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get a lead by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Lead not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateLead
      tags: [leads]
      summary: Update a lead
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeadBody"
      responses:
        "200":
          description: Lead updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Lead not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete a lead
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Lead deleted
        "404":
          description: Lead not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /leads/{id}/status:
    patch:
      operationId: updateLeadStatus
      tags: [leads]
      summary: Update lead status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeadStatusBody"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Lead not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard statistics
      description: Returns summary stats for the CRM dashboard
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /dashboard/recent:
    get:
      operationId: getRecentLeads
      tags: [dashboard]
      summary: Get recent leads activity
      description: Returns the most recently added or updated leads
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Recent leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"

  /dashboard/sources:
    get:
      operationId: getLeadsBySource
      tags: [dashboard]
      summary: Get leads grouped by source
      description: Returns lead counts grouped by source for analytics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Leads by source
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SourceBreakdown"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password

    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/AdminUser"
      required:
        - token
        - user

    AdminUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
      required:
        - id
        - email
        - name

    Lead:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        source:
          type: string
          enum: [Website, Social Media, Referral, Cold Call, Email Campaign, Other]
        status:
          type: string
          enum: [New, Contacted, Converted]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - phone
        - source
        - status
        - createdAt
        - updatedAt

    CreateLeadBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        source:
          type: string
          enum: [Website, Social Media, Referral, Cold Call, Email Campaign, Other]
        status:
          type: string
          enum: [New, Contacted, Converted]
        notes:
          type: ["string", "null"]
      required:
        - name
        - email
        - phone
        - source
        - status

    UpdateLeadBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        source:
          type: string
          enum: [Website, Social Media, Referral, Cold Call, Email Campaign, Other]
        status:
          type: string
          enum: [New, Contacted, Converted]
        notes:
          type: ["string", "null"]

    UpdateLeadStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [New, Contacted, Converted]
      required:
        - status

    LeadsListResponse:
      type: object
      properties:
        leads:
          type: array
          items:
            $ref: "#/components/schemas/Lead"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - leads
        - total
        - page
        - limit
        - totalPages

    DashboardStats:
      type: object
      properties:
        totalLeads:
          type: integer
        newLeads:
          type: integer
        contactedLeads:
          type: integer
        convertedLeads:
          type: integer
        conversionRate:
          type: number
        leadsThisMonth:
          type: integer
      required:
        - totalLeads
        - newLeads
        - contactedLeads
        - convertedLeads
        - conversionRate
        - leadsThisMonth

    SourceBreakdown:
      type: object
      properties:
        source:
          type: string
        count:
          type: integer
      required:
        - source
        - count