# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lead Management CRM API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Authenticate with email and password, returns JWT token
 * @summary Admin login
 */
export const LoginBody = zod.object({
  email: zod.string().email(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  token: zod.string(),
  user: zod.object({
    id: zod.number(),
    email: zod.string(),
    name: zod.string(),
  }),
});

/**
 * Returns authenticated user info
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  email: zod.string(),
  name: zod.string(),
});

/**
 * Returns a list of leads with optional search and filter
 * @summary List all leads
 */
export const GetLeadsQueryParams = zod.object({
  search: zod.coerce.string().optional().describe("Search by name or email"),
  status: zod
    .enum(["New", "Contacted", "Converted"])
    .optional()
    .describe("Filter by status"),
  sort: zod.enum(["newest", "oldest"]).optional().describe("Sort by date"),
  page: zod.coerce.number().optional().describe("Page number"),
  limit: zod.coerce.number().optional().describe("Items per page"),
});

export const GetLeadsResponse = zod.object({
  leads: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      email: zod.string(),
      phone: zod.string(),
      source: zod.enum([
        "Website",
        "Social Media",
        "Referral",
        "Cold Call",
        "Email Campaign",
        "Other",
      ]),
      status: zod.enum(["New", "Contacted", "Converted"]),
      notes: zod.string().nullish(),
      createdAt: zod.coerce.date(),
      updatedAt: zod.coerce.date(),
    }),
  ),
  total: zod.number(),
  page: zod.number(),
  limit: zod.number(),
  totalPages: zod.number(),
});

/**
 * @summary Create a new lead
 */
export const CreateLeadBody = zod.object({
  name: zod.string(),
  email: zod.string().email(),
  phone: zod.string(),
  source: zod.enum([
    "Website",
    "Social Media",
    "Referral",
    "Cold Call",
    "Email Campaign",
    "Other",
  ]),
  status: zod.enum(["New", "Contacted", "Converted"]),
  notes: zod.string().nullish(),
});

/**
 * @summary Get a lead by ID
 */
export const GetLeadParams = zod.object({
  id: zod.coerce.number(),
});

export const GetLeadResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  phone: zod.string(),
  source: zod.enum([
    "Website",
    "Social Media",
    "Referral",
    "Cold Call",
    "Email Campaign",
    "Other",
  ]),
  status: zod.enum(["New", "Contacted", "Converted"]),
  notes: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Update a lead
 */
export const UpdateLeadParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateLeadBody = zod.object({
  name: zod.string().optional(),
  email: zod.string().email().optional(),
  phone: zod.string().optional(),
  source: zod
    .enum([
      "Website",
      "Social Media",
      "Referral",
      "Cold Call",
      "Email Campaign",
      "Other",
    ])
    .optional(),
  status: zod.enum(["New", "Contacted", "Converted"]).optional(),
  notes: zod.string().nullish(),
});

export const UpdateLeadResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  phone: zod.string(),
  source: zod.enum([
    "Website",
    "Social Media",
    "Referral",
    "Cold Call",
    "Email Campaign",
    "Other",
  ]),
  status: zod.enum(["New", "Contacted", "Converted"]),
  notes: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete a lead
 */
export const DeleteLeadParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Update lead status
 */
export const UpdateLeadStatusParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateLeadStatusBody = zod.object({
  status: zod.enum(["New", "Contacted", "Converted"]),
});

export const UpdateLeadStatusResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  phone: zod.string(),
  source: zod.enum([
    "Website",
    "Social Media",
    "Referral",
    "Cold Call",
    "Email Campaign",
    "Other",
  ]),
  status: zod.enum(["New", "Contacted", "Converted"]),
  notes: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * Returns summary stats for the CRM dashboard
 * @summary Get dashboard statistics
 */
export const GetDashboardStatsResponse = zod.object({
  totalLeads: zod.number(),
  newLeads: zod.number(),
  contactedLeads: zod.number(),
  convertedLeads: zod.number(),
  conversionRate: zod.number(),
  leadsThisMonth: zod.number(),
});

/**
 * Returns the most recently added or updated leads
 * @summary Get recent leads activity
 */
export const GetRecentLeadsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  phone: zod.string(),
  source: zod.enum([
    "Website",
    "Social Media",
    "Referral",
    "Cold Call",
    "Email Campaign",
    "Other",
  ]),
  status: zod.enum(["New", "Contacted", "Converted"]),
  notes: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const GetRecentLeadsResponse = zod.array(GetRecentLeadsResponseItem);

/**
 * Returns lead counts grouped by source for analytics
 * @summary Get leads grouped by source
 */
export const GetLeadsBySourceResponseItem = zod.object({
  source: zod.string(),
  count: zod.number(),
});
export const GetLeadsBySourceResponse = zod.array(GetLeadsBySourceResponseItem);