# lib/api-zod/src/generated/types/createLeadBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lead Management CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateLeadBodySource } from "./createLeadBodySource";
import type { CreateLeadBodyStatus } from "./createLeadBodyStatus";

export interface CreateLeadBody {
  name: string;
  email: string;
  phone: string;
  source: CreateLeadBodySource;
  status: CreateLeadBodyStatus;
  /** @nullable */
  notes?: string | null;
}