# lib/api-zod/src/generated/types/updateLeadBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lead Management CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateLeadBodySource } from "./updateLeadBodySource";
import type { UpdateLeadBodyStatus } from "./updateLeadBodyStatus";

export interface UpdateLeadBody {
  name?: string;
  email?: string;
  phone?: string;
  source?: UpdateLeadBodySource;
  status?: UpdateLeadBodyStatus;
  /** @nullable */
  notes?: string | null;
}