# lib/api-zod/src/generated/types/updateLeadStatusBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lead Management CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateLeadStatusBodyStatus } from "./updateLeadStatusBodyStatus";

export interface UpdateLeadStatusBody {
  status: UpdateLeadStatusBodyStatus;
}